transfered the fake data into the story file
