+++ meteor-app/imports/ui/assessment2020/service.js
@@ -1,0 +1,5 @@
+import React from 'react'
+
+export default function Service() {
+  return <div></div>
+}
